+++ napari/util/misc.py
@@ -274,3 +274,78 @@ camel_to_snake_pattern = re.compile(r'(.)([A-Z][a-z]+)')
 def camel_to_snake(name):
     # https://gist.github.com/jaytaylor/3660565
     return camel_to_snake_pattern.sub(r'\1_\2', name).lower()
+
+
+class CallDefault(inspect.Parameter):
+    def __str__(self):
+        """wrap defaults"""
+        kind = self.kind
+        formatted = self._name
+
+        # Fill in defaults
+        if self._default is not inspect._empty:
+            formatted = '{}={}'.format(formatted, formatted)
+
+        if kind == inspect._VAR_POSITIONAL:
+            formatted = '*' + formatted
+        elif kind == inspect._VAR_KEYWORD:
+            formatted = '**' + formatted
+
+        return formatted
+
+
+class CallSignature(inspect.Signature):
+    _parameter_cls = CallDefault
+
+    def __str__(self):
+        """do not render separators
+
+        commented code is what was taken out from
+        the copy/pasted inspect module code :)
+        """
+        result = []
+        # render_pos_only_separator = False
+        # render_kw_only_separator = True
+        for param in self.parameters.values():
+            formatted = str(param)
+
+            # kind = param.kind
+
+            # if kind == inspect._POSITIONAL_ONLY:
+            #     render_pos_only_separator = True
+            # elif render_pos_only_separator:
+            #     # It's not a positional-only parameter, and the flag
+            #     # is set to 'True' (there were pos-only params before.)
+            #     result.append('/')
+            #     render_pos_only_separator = False
+
+            # if kind == inspect._VAR_POSITIONAL:
+            #     # OK, we have an '*args'-like parameter, so we won't need
+            #     # a '*' to separate keyword-only arguments
+            #     render_kw_only_separator = False
+            # elif kind == inspect._KEYWORD_ONLY and render_kw_only_separator:
+            #     # We have a keyword-only parameter to render and we haven't
+            #     # rendered an '*args'-like parameter before, so add a '*'
+            #     # separator to the parameters list ("foo(arg1, *, arg2)" case)
+            #     result.append('*')
+            #     # This condition should be only triggered once, so
+            #     # reset the flag
+            #     render_kw_only_separator = False
+
+            result.append(formatted)
+
+        # if render_pos_only_separator:
+        #     # There were only positional-only parameters, hence the
+        #     # flag was not reset to 'False'
+        #     result.append('/')
+
+        rendered = '({})'.format(', '.join(result))
+
+        if self.return_annotation is not inspect._empty:
+            anno = inspect.formatannotation(self.return_annotation)
+            rendered += ' -> {}'.format(anno)
+
+        return rendered
+
+
+callsignature = CallSignature.from_callable
